Split next oversized slices
