+++ app/src/test/java/com/pocketshell/app/tmux/TmuxSessionViewModelTest.kt
@@ -9,8 +9,6 @@ import com.pocketshell.app.connectivity.TerminalNetworkSnapshot
 import com.pocketshell.app.diagnostics.ReconnectCauseTrail
 import com.pocketshell.app.diagnostics.installRecordingDiagnosticSink
 import com.pocketshell.app.hosts.MainDispatcherRule
-import com.pocketshell.app.projects.FolderListGateway
-import com.pocketshell.core.storage.dao.HostDao
 import com.pocketshell.app.session.AgentConversationRepository
 import com.pocketshell.app.session.AgentConversationSyncStatus
 import com.pocketshell.app.session.ConversationLoadState
@@ -43,7 +41,6 @@ import com.pocketshell.core.terminal.ui.TerminalSurfaceState
 import com.pocketshell.core.tmux.CommandResponse
 import com.pocketshell.core.tmux.TmuxClient
 import com.pocketshell.uikit.model.KeyModifierState
-import com.pocketshell.uikit.model.SessionAgentKind
 import com.pocketshell.core.tmux.TmuxClientException
 import com.pocketshell.core.tmux.TmuxClientFactory
 import com.pocketshell.core.tmux.TmuxDisconnectEvent
@@ -134,462 +131,6 @@ class TmuxSessionViewModelTest : TmuxSessionViewModelTestBase() {
     fun connectionStatusStartsIdle() {
         val vm = newVm()
         assertTrue(vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Idle)
-    }
-
-    @Test
-    fun killCurrentSessionBroadcastsSignalOnConfirmedKill() = runTest(scheduler) {
-        // Issue #655: the in-session Stop now routes through the SAME verified
-        // gateway SSH-exec path the host-detail Stop uses (#518), NOT the
-        // control channel it is attached to. A gateway-confirmed kill (the
-        // session is gone) broadcasts the lifecycle signal so the tree drops
-        // the row.
-        val signals = SessionLifecycleSignals()
-        val gateway = RecordingStopGateway(killSucceeds = true)
-        val vm = newVm(
-            sessionLifecycleSignals = signals,
-            folderListGateway = gateway,
-            hostDao = StopHostDao(hostId = 7L),
-        )
-        val client = FakeTmuxClient()
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "docker",
-            host = "10.0.2.2",
-            port = 2222,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "doomed",
-            client = client,
-        )
-        runCurrent()
-
-        val killed = async { signals.killedSessions.first() }
-        runCurrent()
-
-        vm.killCurrentSession()
-        advanceUntilIdle()
-
-        val event = killed.await()
-        assertEquals(7L, event.hostId)
-        assertEquals("doomed", event.sessionName)
-        assertTrue(
-            "expected the gateway kill to target 'doomed', got ${gateway.killedSessionNames}",
-            gateway.killedSessionNames.contains("doomed"),
-        )
-        // The Stop must NOT race its own teardown over the control channel.
-        assertFalse(
-            "in-session Stop must not kill over the control channel; sent=${client.sentCommands}",
-            client.sentCommands.any { it.startsWith("kill-session") },
-        )
-    }
-
-    @Test
-    fun confirmedKillEvictsKilledSessionFromNameKeyedCaches() = runTest(scheduler) {
-        val runtimeCache = TmuxSessionRuntimeCache(maxEntries = 4, nowMs = { 0L })
-        val agentSessionMemory = AgentSessionMemory()
-        val signals = SessionLifecycleSignals(
-            runtimeCache = runtimeCache,
-            agentSessionMemory = agentSessionMemory,
-        )
-        val doomedRuntime = cachedRuntimeForTest(sessionName = "doomed", hostId = 7L)
-        val otherRuntime = cachedRuntimeForTest(sessionName = "other", hostId = 7L)
-        runtimeCache.put(doomedRuntime)
-        runtimeCache.put(otherRuntime)
-        agentSessionMemory.remember(
-            hostId = 7L,
-            sessionName = "doomed",
-            windowId = "@2",
-            detection = AgentDetection(
-                agent = AgentKind.ClaudeCode,
-                sourcePath = "/home/alex/.claude/projects/doomed.jsonl",
-                sessionId = "doomed-agent",
-                confidence = AgentDetection.Confidence.ProcessConfirmed,
-            ),
-            wasOnConversation = true,
-        )
-        val gateway = RecordingStopGateway(killSucceeds = true)
-        val vm = newVm(
-            runtimeCache = runtimeCache,
-            agentSessionMemory = agentSessionMemory,
-            sessionLifecycleSignals = signals,
-            folderListGateway = gateway,
-            hostDao = StopHostDao(hostId = 7L),
-        )
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "docker",
-            host = "10.0.2.2",
-            port = 2222,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "doomed",
-            client = FakeTmuxClient(),
-        )
-        runCurrent()
-
-        val killed = async { signals.killedSessions.first() }
-        runCurrent()
-
-        vm.killCurrentSession()
-        advanceUntilIdle()
-
-        val event = killed.await()
-        assertEquals(7L, event.hostId)
-        assertEquals("doomed", event.sessionName)
-        assertFalse(
-            "confirmed Stop must evict the killed session's parked runtime",
-            runtimeCache.contains(doomedRuntime.key),
-        )
-        assertTrue(runtimeCache.contains(otherRuntime.key))
-        assertNull(
-            "same-name successor must not recall agent memory from the killed session",
-            agentSessionMemory.recall(7L, "doomed", "@2"),
-        )
-    }
-
-    // ----- Issue #898: the in-session kebab "+ New session" now opens the SAME
-    // rich SessionTypePickerSheet the host screen uses, and its Create routes
-    // through the SAME verified gateway createSession path so the created
-    // session honours the chosen type / CLI / skip-permissions / profile. These
-    // tests assert that the agent-CLI + skip-permissions + profile choices the
-    // picker synthesises into a `startCommand` reach the gateway intact (the old
-    // name+folder dialog path could carry NONE of these — it only had a
-    // startDirectory). Red→green: on base (the old `createSession(name,
-    // startDirectory)` that sent `new-session -d` over the control channel and
-    // had no startCommand/gateway route), these assertions cannot pass.
-
-    @Test
-    fun createSessionRoutesChosenAgentOptionsThroughGateway() = runTest(scheduler) {
-        val gateway = RecordingStopGateway(
-            killSucceeds = true,
-            createResolvedName = "git-claude",
-        )
-        val vm = newVm(
-            folderListGateway = gateway,
-            hostDao = StopHostDao(hostId = 7L),
-        )
-        val client = FakeTmuxClient()
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "docker",
-            host = "10.0.2.2",
-            port = 2222,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = client,
-        )
-        runCurrent()
-
-        // The rich sheet synthesises this startCommand for an Agent=codex,
-        // skip-permissions OFF, profile "work" choice. We pass it straight to
-        // createSession exactly as TmuxSessionScreen does.
-        val startCommand = "pocketshell agent codex --dir '/home/alex/git' " +
-            "--no-skip-permissions --profile 'work'"
-        // The create routes a `dao.getById` over Dispatchers.IO (real IO, not the
-        // virtual clock), so await onResolved deterministically rather than
-        // racing advanceUntilIdle — the kill tests await a signal for the same
-        // reason. `await()` parks the test body until the real-IO continuation
-        // completes the deferred, then runTest unparks and we assert.
-        val resolvedDeferred = CompletableDeferred<String>()
-        vm.createSession(
-            name = "git-codex",
-            cwd = "/home/alex/git",
-            startCommand = startCommand,
-            chosenKind = SessionAgentKind.Codex,
-            onResolved = { resolvedDeferred.complete(it) },
-        )
-        val resolved = resolvedDeferred.await()
-        advanceUntilIdle()
-
-        assertEquals(
-            "expected exactly one gateway createSession call, got ${gateway.createCalls}",
-            1,
-            gateway.createCalls.size,
-        )
-        val call = gateway.createCalls.single()
-        assertEquals("git-codex", call.sessionName)
-        assertEquals("/home/alex/git", call.cwd)
-        // The load-bearing assertion (G6): the chosen CLI / skip-perms / profile
-        // are all encoded in the startCommand that reached the verified gateway.
-        assertEquals(startCommand, call.startCommand)
-        // onResolved fired with the gateway's resolved name so the screen can
-        // navigate/attach to the freshly-created session.
-        assertEquals("git-claude", resolved)
-        // The rich-sheet create must NOT fall back to the control-channel
-        // `new-session -d` (which cannot launch an agent CLI).
-        assertFalse(
-            "rich-sheet create must use the gateway, not control-channel new-session; " +
-                "sent=${client.sentCommands}",
-            client.sentCommands.any { it.startsWith("new-session") },
-        )
-    }
-
-    @Test
-    fun createSessionShellChoiceHasNoStartCommand() = runTest(scheduler) {
-        val gateway = RecordingStopGateway(
-            killSucceeds = true,
-            createResolvedName = "plain",
-        )
-        val vm = newVm(
-            folderListGateway = gateway,
-            hostDao = StopHostDao(hostId = 7L),
-        )
-        val client = FakeTmuxClient()
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "docker",
-            host = "10.0.2.2",
-            port = 2222,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = client,
-        )
-        runCurrent()
-
-        // A Shell choice synthesises a null startCommand. Await onResolved so the
-        // real-IO `dao.getById` hop completes before we assert (see the agent
-        // test above for why advanceUntilIdle alone races the IO).
-        val resolvedDeferred = CompletableDeferred<String>()
-        vm.createSession(
-            name = "plain",
-            cwd = "/srv/app",
-            startCommand = null,
-            chosenKind = null,
-            onResolved = { resolvedDeferred.complete(it) },
-        )
-        resolvedDeferred.await()
-        advanceUntilIdle()
-
-        val call = gateway.createCalls.single()
-        assertEquals("/srv/app", call.cwd)
-        assertNull("shell session must carry no startCommand", call.startCommand)
-    }
-
-    @Test
-    fun killCurrentSessionDoesNotBroadcastWhenGatewayKillFails() = runTest(scheduler) {
-        // Issue #655: a gateway-verified failure (the session is STILL running
-        // after the kill, e.g. `tmux has-session` still exits 0) must NOT
-        // broadcast, so the tree keeps the still-live row.
-        val signals = SessionLifecycleSignals()
-        val gateway = RecordingStopGateway(killSucceeds = false)
-        val vm = newVm(
-            sessionLifecycleSignals = signals,
-            folderListGateway = gateway,
-            hostDao = StopHostDao(hostId = 7L),
-        )
-        val client = FakeTmuxClient()
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "docker",
-            host = "10.0.2.2",
-            port = 2222,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "stubborn",
-            client = client,
-        )
-        runCurrent()
-
-        var broadcast: KilledSession? = null
-        val collector = async { broadcast = signals.killedSessions.first() }
-        runCurrent()
-
-        vm.killCurrentSession()
-        advanceUntilIdle()
-
-        assertNull("a failed (unverified) kill must not broadcast a lifecycle signal", broadcast)
-        collector.cancel()
-    }
-
-    // ----------------------------------------------------------- Issue #883
-
-    /**
-     * Issue #883 (reproduce-first). The tree presents each tmux WINDOW as its
-     * own `[wN]` row, but "Stop session" used to ALWAYS run `kill-session`,
-     * taking the whole session (every window) down. This proves the fix: Stop
-     * on a window row of a MULTI-window session runs `kill-window` for ONLY
-     * that window — NOT `kill-session` — and, because a sibling window
-     * survived, broadcasts a WINDOW-close (not a whole-session kill) so the
-     * tree drops only the killed window row.
-     *
-     * Pre-fix `killCurrentSession(windowIndex)` ignored the index and called
-     * `killSession`, so `killedSessionNames` would contain "multi" and
-     * `killedWindowTargets` would be empty — this test fails on base.
-     */
-    @Test
-    fun stopOnMultiWindowRowKillsOnlyThatWindow() = runTest(scheduler) {
-        val signals = SessionLifecycleSignals()
-        val gateway = RecordingStopGateway(
-            killSucceeds = true,
-            windowKillSessionSurvived = true,
-        )
-        val vm = newVm(
-            sessionLifecycleSignals = signals,
-            folderListGateway = gateway,
-            hostDao = StopHostDao(hostId = 7L),
-        )
-        val client = FakeTmuxClient()
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "docker",
-            host = "10.0.2.2",
-            port = 2222,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "multi",
-            client = client,
-        )
-        runCurrent()
-        // Two windows of the same session — `[w0]` (@10) and `[w1]` (@11).
-        vm.applyParsedPanesForTest(
-            listOf(
-                TmuxSessionViewModel.ParsedPane(
-                    "%0", "@10", "$0", "win0", paneIndex = 0,
-                    windowIndex = 0, sessionName = "multi",
-                ),
-                TmuxSessionViewModel.ParsedPane(
-                    "%1", "@11", "$0", "win1", paneIndex = 0,
-                    windowIndex = 1, sessionName = "multi",
-                ),
-            ),
-        )
-        advanceUntilIdle()
-
-        val closed = async { signals.closedWindows.first() }
-        runCurrent()
-
-        // The user is viewing window 0 (`[w0]`) and taps Stop.
-        vm.killCurrentSession(windowIndex = 0)
-        advanceUntilIdle()
-
-        val event = closed.await()
-        assertEquals(7L, event.hostId)
-        assertEquals(
-            "the closed-window signal must carry window 0's stable tmux id (@10)",
-            "@10",
-            event.windowId,
-        )
-        assertEquals(
-            "Stop on a window row must run kill-window for ONLY that window; " +
-                "got ${gateway.killedWindowTargets}",
-            listOf("multi:0"),
-            gateway.killedWindowTargets,
-        )
-        assertTrue(
-            "Stop on a window row must NOT run kill-session (that would take " +
-                "the whole session + sibling window down); got ${gateway.killedSessionNames}",
-            gateway.killedSessionNames.isEmpty(),
-        )
-    }
-
-    /**
-     * Issue #883 class coverage: Stop on the LAST/only window of a session.
-     * tmux auto-destroys the session when its final window closes, so the
-     * gateway reports `sessionSurvived = false` and the VM must broadcast a
-     * whole-session kill (the row drops entirely), exactly as today.
-     */
-    @Test
-    fun stopOnLastWindowDestroysTheSession() = runTest(scheduler) {
-        val signals = SessionLifecycleSignals()
-        val gateway = RecordingStopGateway(
-            killSucceeds = true,
-            windowKillSessionSurvived = false,
-        )
-        val vm = newVm(
-            sessionLifecycleSignals = signals,
-            folderListGateway = gateway,
-            hostDao = StopHostDao(hostId = 7L),
-        )
-        val client = FakeTmuxClient()
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "docker",
-            host = "10.0.2.2",
-            port = 2222,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "solo",
-            client = client,
-        )
-        runCurrent()
-        vm.applyParsedPanesForTest(
-            listOf(
-                TmuxSessionViewModel.ParsedPane(
-                    "%0", "@20", "$0", "win0", paneIndex = 0,
-                    windowIndex = 0, sessionName = "solo",
-                ),
-            ),
-        )
-        advanceUntilIdle()
-
-        val killed = async { signals.killedSessions.first() }
-        runCurrent()
-
-        vm.killCurrentSession(windowIndex = 0)
-        advanceUntilIdle()
-
-        val event = killed.await()
-        assertEquals(7L, event.hostId)
-        assertEquals(
-            "closing the last window destroys the session — the whole-session " +
-                "signal must carry the session name",
-            "solo",
-            event.sessionName,
-        )
-        assertEquals(
-            "the last-window Stop still goes through kill-window on the remote",
-            listOf("solo:0"),
-            gateway.killedWindowTargets,
-        )
-    }
-
-    /**
-     * Issue #883 class coverage: the whole-session Stop path (no window index)
-     * still runs `kill-session` and broadcasts a session kill — the session-
-     * level / fallback intent is unchanged.
-     */
-    @Test
-    fun stopWithoutWindowIndexKillsWholeSession() = runTest(scheduler) {
-        val signals = SessionLifecycleSignals()
-        val gateway = RecordingStopGateway(killSucceeds = true)
-        val vm = newVm(
-            sessionLifecycleSignals = signals,
-            folderListGateway = gateway,
-            hostDao = StopHostDao(hostId = 7L),
-        )
-        val client = FakeTmuxClient()
-        vm.replaceClientForTest(
-            hostId = 7L,
-            hostName = "docker",
-            host = "10.0.2.2",
-            port = 2222,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "whole",
-            client = client,
-        )
-        runCurrent()
-
-        val killed = async { signals.killedSessions.first() }
-        runCurrent()
-
-        // No window index — the session-level Stop.
-        vm.killCurrentSession()
-        advanceUntilIdle()
-
-        val event = killed.await()
-        assertEquals("whole", event.sessionName)
-        assertEquals(
-            "the no-window Stop must run kill-session, not kill-window",
-            listOf("whole"),
-            gateway.killedSessionNames,
-        )
-        assertTrue(
-            "the no-window Stop must NOT run kill-window; got ${gateway.killedWindowTargets}",
-            gateway.killedWindowTargets.isEmpty(),
-        )
     }
 
     // ----------------------------------------------------------- Issue #665
@@ -16851,141 +16392,6 @@ class TmuxSessionViewModelTest : TmuxSessionViewModelTestBase() {
         ): String = error("uploadStream not used in this test")
 
         override fun close() = Unit
-    }
-
-    /**
-     * Issue #655: minimal [FolderListGateway] that records the session names
-     * `killSession` was asked to stop. Drives the in-session Stop's unified
-     * verified-kill path without a live SSH/tmux server. [killSucceeds]
-     * toggles the gateway's confirmed-gone vs still-running outcome so the
-     * "broadcast only on confirmed kill" contract can be asserted both ways.
-     */
-    private class RecordingStopGateway(
-        private val killSucceeds: Boolean,
-        // Issue #883: window-kill outcome. `null` means killWindow was never
-        // expected to be called (a session-level kill test). `true` => a
-        // sibling window survived; `false` => the session was destroyed.
-        private val windowKillSessionSurvived: Boolean? = null,
-        private val windowKillSucceeds: Boolean = true,
-        // Issue #898: when set, createSession succeeds returning this resolved
-        // name; otherwise it is "not used" (the kill tests never create).
-        private val createResolvedName: String? = null,
-    ) : FolderListGateway {
-        val killedSessionNames = mutableListOf<String>()
-
-        // Issue #883: records the `<sessionName>:<windowIndex>` window targets
-        // killWindow was asked to stop, plus whether kill-session was used.
-        val killedWindowTargets = mutableListOf<String>()
-
-        // Issue #898: the exact (sessionName, cwd, startCommand) the in-session
-        // "+ New session" rich-sheet create path passed through to the gateway —
-        // the verified host-screen path. The startCommand encodes the chosen
-        // CLI / skip-permissions / profile, so asserting on it proves every
-        // picked option is honoured end-to-end.
-        data class CreateCall(
-            val sessionName: String,
-            val cwd: String,
-            val startCommand: String?,
-        )
-        val createCalls = mutableListOf<CreateCall>()
-
-        override suspend fun listSessionsWithFolder(
-            host: com.pocketshell.core.storage.entity.HostEntity,
-            keyPath: String,
-            passphrase: CharArray?,
-            watchedRoots: List<com.pocketshell.core.storage.entity.ProjectRootEntity>,
-        ): com.pocketshell.app.projects.FolderListResult =
-            com.pocketshell.app.projects.FolderListResult.Sessions(rows = emptyList())
-
-        override suspend fun createSession(
-            host: com.pocketshell.core.storage.entity.HostEntity,
-            keyPath: String,
-            passphrase: CharArray?,
-            sessionName: String,
-            cwd: String,
-            startCommand: String?,
-        ): Result<String> {
-            createCalls += CreateCall(sessionName, cwd, startCommand)
-            return createResolvedName?.let { Result.success(it) } ?: error("not used")
-        }
-
-        override suspend fun createEmptyProject(
-            host: com.pocketshell.core.storage.entity.HostEntity,
-            keyPath: String,
-            passphrase: CharArray?,
-            parentPath: String,
-            folderName: String,
-        ): Result<String> = error("not used")
-
-        override suspend fun importFile(
-            host: com.pocketshell.core.storage.entity.HostEntity,
-            keyPath: String,
-            passphrase: CharArray?,
-            folderPath: String,
-            payload: com.pocketshell.app.projects.FolderImportPayload,
-        ): Result<String> = error("not used")
-
-        override suspend fun killSession(
-            host: com.pocketshell.core.storage.entity.HostEntity,
-            keyPath: String,
-            passphrase: CharArray?,
-            sessionName: String,
-        ): Result<Unit> {
-            if (!killSucceeds) {
-                return Result.failure(
-                    RuntimeException("tmux session '$sessionName' is still running."),
-                )
-            }
-            killedSessionNames += sessionName
-            return Result.success(Unit)
-        }
-
-        override suspend fun killWindow(
-            host: com.pocketshell.core.storage.entity.HostEntity,
-            keyPath: String,
-            passphrase: CharArray?,
-            sessionName: String,
-            windowIndex: Int,
-        ): Result<com.pocketshell.app.projects.WindowKillOutcome> {
-            killedWindowTargets += "$sessionName:$windowIndex"
-            if (!windowKillSucceeds) {
-                return Result.failure(
-                    RuntimeException("tmux window '$sessionName:$windowIndex' is still running."),
-                )
-            }
-            return Result.success(
-                com.pocketshell.app.projects.WindowKillOutcome(
-                    sessionSurvived = windowKillSessionSurvived ?: false,
-                ),
-            )
-        }
-    }
-
-    /**
-     * Issue #655: host DAO that returns one host for [hostId] so the
-     * in-session Stop can resolve the [HostEntity] it hands the gateway.
-     */
-    private class StopHostDao(private val hostId: Long) : HostDao {
-        private val host = com.pocketshell.core.storage.entity.HostEntity(
-            id = hostId,
-            name = "docker",
-            hostname = "10.0.2.2",
-            port = 2222,
-            username = "alex",
-            keyId = 1L,
-        )
-
-        override fun getAll() = kotlinx.coroutines.flow.flowOf(listOf(host))
-        override suspend fun getById(id: Long): com.pocketshell.core.storage.entity.HostEntity? =
-            host.takeIf { it.id == id }
-        override fun getEnabled() = kotlinx.coroutines.flow.flowOf(listOf(host))
-        override suspend fun insert(host: com.pocketshell.core.storage.entity.HostEntity): Long =
-            error("not used")
-        override suspend fun update(host: com.pocketshell.core.storage.entity.HostEntity) =
-            error("not used")
-        override suspend fun delete(host: com.pocketshell.core.storage.entity.HostEntity) =
-            error("not used")
-        override suspend fun deleteById(id: Long) = error("not used")
     }
 
     private class OutputFailingTmuxClient(
