+++ scripts/file-size-hygiene-baseline.txt
@@ -9,13 +9,11 @@
 139372 app/src/main/java/com/pocketshell/app/projects/FolderListScreen.kt
 180197 app/src/main/java/com/pocketshell/app/projects/FolderListViewModel.kt
 226939 app/src/main/java/com/pocketshell/app/tmux/TmuxSessionScreen.kt
-950083 app/src/main/java/com/pocketshell/app/tmux/TmuxSessionViewModel.kt
+946455 app/src/main/java/com/pocketshell/app/tmux/TmuxSessionViewModel.kt
 278779 app/src/test/java/com/pocketshell/app/composer/PromptComposerViewModelTest.kt
-138541 app/src/test/java/com/pocketshell/app/session/AgentConversationRepositoryTest.kt
-728066 app/src/test/java/com/pocketshell/app/tmux/TmuxSessionViewModelTest.kt
+704615 app/src/test/java/com/pocketshell/app/tmux/TmuxSessionViewModelTest.kt
 1146350 docs/mockups/feedback/folder-tree-target-20260604.png
 182790 docs/screenshots/readme-settings.png
 194363 scripts/ci-journey-suite.sh
 160115 shared/core-tmux/src/main/java/com/pocketshell/core/tmux/TmuxClient.kt
-152167 shared/core-tmux/src/test/java/com/pocketshell/core/tmux/TmuxClientTest.kt
-144031 shared/ui-kit/src/test/java/com/pocketshell/uikit/render/DesignRenders.kt
+138387 shared/ui-kit/src/test/java/com/pocketshell/uikit/render/DesignRenders.kt
