+++ shared/core-tmux/src/test/java/com/pocketshell/core/tmux/TmuxClientTest.kt
@@ -18,7 +18,6 @@ import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.take
-import kotlinx.coroutines.launch
 import kotlinx.coroutines.flow.toList
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.selects.select
@@ -42,7 +41,6 @@ import java.nio.charset.StandardCharsets
 import java.util.Collections
 import java.util.concurrent.CountDownLatch
 import java.util.concurrent.TimeUnit
-import java.util.concurrent.atomic.AtomicInteger
 
 /**
  * Generous ceiling for await-style waits (`withTimeout(N) { deferred.await() }`
@@ -277,221 +275,6 @@ class TmuxClientTest {
             client.close()
         }
     }
-
-    @Test
-    fun `pauseOutputDelivery holds output across a collector gap so a fresh collector recovers them`() =
-        runBlocking {
-            // Issue #1297 (acceptance criterion 2, red→green): the overflow-reseed
-            // producer swap tears the sole pane collector down and reattaches a
-            // fresh one. On BASE the %output emitted in that gap is emitted into
-            // the zero-subscriber (replay = 0) flow and lost, so a re-subscribing
-            // collector recovers NOTHING. With the fix, pausing delivery across
-            // the swap HOLDS the frames in the bounded backlog and they replay to
-            // the fresh collector in order on resume.
-            val shell = FakeShell()
-            val session = FakeSession(shell)
-            val client = RealTmuxClient(session, scope)
-            try {
-                client.connect()
-                // Attach the sole collector so the pane pipe exists and is live.
-                val firstCollector = scope.launch { client.outputFor("%7").collect { } }
-                delay(150)
-                // Freeze delivery, THEN detach the sole collector (mirrors the
-                // overflow reseed order: pause before the producer teardown).
-                client.pauseOutputDelivery("%7")
-                firstCollector.cancelAndJoin()
-
-                // Emit N frames into the collector gap.
-                shell.feed(
-                    "%output %7 gap-one\n" +
-                        "%output %7 gap-two\n" +
-                        "%output %7 gap-three\n",
-                )
-                // Let the reader parse them into the (paused) backlog.
-                delay(250)
-
-                // Re-subscribe a fresh collector, then thaw — held frames drain in
-                // arrival order to the fresh collector.
-                val recovered = scope.async { client.outputFor("%7").take(3).toList() }
-                delay(100)
-                client.resumeOutputDelivery("%7")
-
-                val frames = withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { recovered.await() }
-                assertEquals(
-                    "every %output emitted during the collector gap must replay in order",
-                    listOf("gap-one", "gap-two", "gap-three"),
-                    frames.map { String(it.data, StandardCharsets.US_ASCII) },
-                )
-            } finally {
-                client.close()
-            }
-        }
-
-    @Test
-    fun `output emitted during an unpaused collector gap is dropped`() = runBlocking {
-        // Issue #1297: documents the SPOF the pause closes. WITHOUT the delivery
-        // pause, %output emitted while the sole collector is detached vanishes
-        // into the zero-subscriber flow — a re-subscribing collector recovers
-        // none of it. This is the base behaviour the overflow-reseed swap relied
-        // on the capture to paper over; `pauseOutputDelivery` is the opt-in fix.
-        val shell = FakeShell()
-        val session = FakeSession(shell)
-        val client = RealTmuxClient(session, scope)
-        try {
-            client.connect()
-            val firstCollector = scope.launch { client.outputFor("%8").collect { } }
-            delay(150)
-            // Detach WITHOUT pausing.
-            firstCollector.cancelAndJoin()
-            shell.feed(
-                "%output %8 lost-one\n" +
-                    "%output %8 lost-two\n",
-            )
-            delay(250)
-            // Re-subscribe and prove the gap frames are gone: a fresh sentinel is
-            // the only thing delivered.
-            val recovered = scope.async { client.outputFor("%8").take(1).toList() }
-            delay(100)
-            shell.feed("%output %8 after-resubscribe\n")
-            val frames = withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { recovered.await() }
-            assertEquals(
-                "unpaused gap frames are dropped; only post-resubscribe output survives",
-                listOf("after-resubscribe"),
-                frames.map { String(it.data, StandardCharsets.US_ASCII) },
-            )
-        } finally {
-            client.close()
-        }
-    }
-
-    @Test
-    fun `capture SUCCESS drains the frame parked at the pause boundary, not double-applied after the snapshot`() =
-        runBlocking {
-            // Issue #1305 (red→green, SUCCESS path): the overflow-reseed swap pauses
-            // %output, drains the pane backlog (step 2, PRE-capture), takes an
-            // authoritative capture snapshot, drains AGAIN (step 4, POST-capture),
-            // then resumes. Both drains empty the pipe CHANNEL — but NOT the single
-            // frame already received into the drain loop's local and PARKED at the
-            // pause boundary (`if (paused.value) { paused.first { !it } }`). On the
-            // #1297 base that parked frame survives BOTH drains and is emitted on
-            // resume, AFTER the snapshot that already reflects it — a double-apply on
-            // top of the fresh grid. With the fix the AUTHORITATIVE step-4 drain
-            // (`drainPaneOutputBacklogAfterCapture`) marks the parked frame stale so
-            // it is dropped on resume; only genuinely post-snapshot deltas reach the
-            // collector.
-            val shell = FakeShell()
-            val session = FakeSession(shell)
-            val client = RealTmuxClient(session, scope)
-            val received = Collections.synchronizedList(mutableListOf<String>())
-            try {
-                client.connect()
-                // Live collector so the pane pipe's drain loop is hot (blocked in
-                // receiveCatching on the empty channel, ready to receive + park).
-                val collector = scope.launch {
-                    client.outputFor("%305").collect {
-                        received.add(String(it.data, StandardCharsets.US_ASCII))
-                    }
-                }
-                delay(200)
-
-                // 1. Pause BEFORE feeding: the next frame the loop receives lands
-                //    in its local and PARKS at the pause boundary (it is NOT left
-                //    queued in the channel).
-                client.pauseOutputDelivery("%305")
-                shell.feed("%output %305 parked-pre-capture\n")
-                // Let the reader parse the frame and the hot loop receive it into
-                // its local + suspend on `paused.first { !it }`.
-                delay(300)
-
-                // 2. Step-2 PRE-capture drain (plain — must NOT arm discard). The
-                //    parked frame is in the loop's LOCAL, not the channel.
-                val drainedPreCapture = client.drainPaneOutputBacklog("%305")
-
-                // 3. Step-4 POST-capture AUTHORITATIVE drain — a successful reseed
-                //    ran, so this arms discard of the parked pre-capture frame.
-                val drainedPostCapture = client.drainPaneOutputBacklogAfterCapture("%305")
-
-                // 4. Resume (snapshot is authoritative). The parked, pre-capture
-                //    frame must NOT reach the collector.
-                client.resumeOutputDelivery("%305")
-                delay(150)
-
-                // 5. A genuinely post-snapshot live delta — this MUST be delivered.
-                shell.feed("%output %305 post-snapshot-delta\n")
-                delay(300)
-
-                collector.cancelAndJoin()
-
-                assertEquals(
-                    "on a SUCCESSFUL reseed the frame parked at the pause boundary " +
-                        "must be dropped by the authoritative post-capture drain (not " +
-                        "double-applied on top of the snapshot); only the post-snapshot " +
-                        "delta survives. drainedPreCapture=" + drainedPreCapture +
-                        " drainedPostCapture=" + drainedPostCapture,
-                    listOf("post-snapshot-delta"),
-                    received.toList(),
-                )
-            } finally {
-                client.close()
-            }
-        }
-
-    @Test
-    fun `capture FAILURE replays the frame parked at the pause boundary, not dropped (issue 1297 guarantee)`() =
-        runBlocking {
-            // Issue #1305 (red→green, FAILURE path — the reviewer's G2 gap): on a
-            // capture FAILURE the overflow-reseed swap runs ONLY the step-2
-            // PRE-capture drain (`drainPaneOutputBacklog`); the step-4 authoritative
-            // drain is guarded by `if (reseeded)` and is SKIPPED. The #1297
-            // held-frame guarantee is that on that failed path the held frames
-            // (including the one parked in the drain loop's local at the pause
-            // boundary) are REPLAYED to the fresh producer on resume, not lost.
-            //
-            // The reviewer flagged that arming discard on ANY drain (the prior fix)
-            // would DROP the parked frame on this failed path too, reintroducing the
-            // #1297 lost-frame break. This test guards that: with the pre-capture
-            // drain NOT arming discard, the parked frame must survive to resume.
-            val shell = FakeShell()
-            val session = FakeSession(shell)
-            val client = RealTmuxClient(session, scope)
-            val received = Collections.synchronizedList(mutableListOf<String>())
-            try {
-                client.connect()
-                val collector = scope.launch {
-                    client.outputFor("%305").collect {
-                        received.add(String(it.data, StandardCharsets.US_ASCII))
-                    }
-                }
-                delay(200)
-
-                // 1. Pause, feed → the frame parks in the drain loop's local.
-                client.pauseOutputDelivery("%305")
-                shell.feed("%output %305 parked-pre-capture\n")
-                delay(300)
-
-                // 2. Step-2 PRE-capture drain — the ONLY drain on the FAILURE path
-                //    (no reseed → step 4 is skipped). It must NOT arm discard.
-                val drainedPreCapture = client.drainPaneOutputBacklog("%305")
-
-                // 3. Resume — capture failed, so the held/parked frame must REPLAY
-                //    to the fresh producer (the #1297 guarantee), not be dropped.
-                client.resumeOutputDelivery("%305")
-                delay(300)
-
-                collector.cancelAndJoin()
-
-                assertEquals(
-                    "on a FAILED reseed (only the pre-capture drain ran, step 4 " +
-                        "skipped) the frame parked at the pause boundary must be " +
-                        "REPLAYED on resume — not silently dropped — preserving the " +
-                        "#1297 held-frame guarantee. drainedPreCapture=" + drainedPreCapture,
-                    listOf("parked-pre-capture"),
-                    received.toList(),
-                )
-            } finally {
-                client.close()
-            }
-        }
 
     @Test
     fun `sendChainedCommands writes one chained line and drains both blocks in order`() = runBlocking {
@@ -1040,835 +823,6 @@ class TmuxClientTest {
             client.close()
         }
     }
-
-    @Test
-    fun `outputFor demuxes per pane id`() = runBlocking {
-        val shell = FakeShell()
-        val session = FakeSession(shell)
-        val client = RealTmuxClient(session, scope)
-        try {
-            client.connect()
-            val pane0Events = scope.async {
-                client.outputFor("%0").take(2).toList()
-            }
-            val pane1Events = scope.async {
-                client.outputFor("%1").take(1).toList()
-            }
-            delay(100) // let subscribers attach to the shared flow
-
-            shell.feed(
-                "%output %0 hello\n" +
-                    "%output %1 world\n" +
-                    "%output %0 again\n",
-            )
-
-            val p0 = withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { pane0Events.await() }
-            val p1 = withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { pane1Events.await() }
-
-            assertEquals(2, p0.size)
-            assertEquals("%0", p0[0].paneId)
-            assertEquals("hello", String(p0[0].data, StandardCharsets.US_ASCII))
-            assertEquals("again", String(p0[1].data, StandardCharsets.US_ASCII))
-
-            assertEquals(1, p1.size)
-            assertEquals("%1", p1[0].paneId)
-            assertEquals("world", String(p1[0].data, StandardCharsets.US_ASCII))
-        } finally {
-            client.close()
-        }
-    }
-
-    @Test
-    fun `outputFor delivers pane output while the structural event collector is parked`() = runBlocking {
-        val shell = FakeShell()
-        val session = FakeSession(shell)
-        val client = RealTmuxClient(session, scope)
-        val firstGlobalEventSeen = CompletableDeferred<Unit>()
-        val releaseGlobalCollector = CompletableDeferred<Unit>()
-        try {
-            client.connect()
-
-            // A parked structural-event collector — models a GC-stalled ViewModel
-            // subscriber on the events bus. Issue #1224: `%output` is no longer
-            // copied onto this bus, so a stalled collector here can NEVER starve
-            // the per-pane render path ([outputFor]).
-            val globalCollector = scope.async {
-                client.events.collect {
-                    firstGlobalEventSeen.complete(Unit)
-                    releaseGlobalCollector.await()
-                }
-            }
-            val targetOutput = scope.async {
-                client.outputFor("%target").first()
-            }
-            delay(100) // let both subscribers attach to their hot flows
-
-            val feedJob = scope.async {
-                shell.feed(
-                    buildString {
-                        // One structural event parks the collector, then a heavy
-                        // output flood well past EVENT_BUFFER (256). Post-#1224 the
-                        // flood never touches the events bus, so the parked
-                        // collector cannot stall %target's per-pane delivery.
-                        append("%session-changed \$0 main\n")
-                        repeat(257) { index ->
-                            append("%output %noise noisy-")
-                            append(index)
-                            append('\n')
-                        }
-                        append("%output %target visible\n")
-                    },
-                )
-            }
-
-            withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { firstGlobalEventSeen.await() }
-            val output = withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { targetOutput.await() }
-
-            assertEquals("%target", output.paneId)
-            assertEquals("visible", String(output.data, StandardCharsets.US_ASCII))
-
-            releaseGlobalCollector.complete(Unit)
-            withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { feedJob.await() }
-            globalCollector.cancel()
-        } finally {
-            releaseGlobalCollector.complete(Unit)
-            client.close()
-        }
-    }
-
-    /**
-     * Issue #1224 (regression, red→green core). Structural events must NOT be
-     * droppable by high-rate `%output` volume.
-     *
-     * Before #1224 every `%output` was copied onto the shared 256-slot
-     * [RealTmuxClient.events] bus purely so the ViewModel could log first-frame.
-     * A dense output burst behind a briefly-stalled collector filled that buffer,
-     * so a `%window-close` at the burst tail was silently dropped ([tryEmit]
-     * returns false when full) — leaving a stale window node in the tree until
-     * some later structural event happened to re-trigger a reconcile.
-     *
-     * Here a structural-event collector is parked for the whole flood while 1000
-     * `%output` frames (≫ the 256 buffer) are fed followed by a trailing
-     * `%window-close @7`. A per-pane sentinel (delivered via [outputFor], which
-     * bypasses the events bus on BOTH base and fix) is the reader-progress
-     * barrier. On base the window-close was dropped → the collector never sees it
-     * (RED). With the fix `%output` is off the bus entirely, so the window-close
-     * is the only thing on it and is delivered (GREEN).
-     */
-    @Test
-    fun `trailing window-close survives an output flood behind a stalled collector`() = runBlocking {
-        val shell = FakeShell()
-        val session = FakeSession(shell)
-        val client = RealTmuxClient(session, scope)
-        val windowCloseSeen = CompletableDeferred<String>()
-        val floodDrained = CompletableDeferred<Unit>()
-        try {
-            client.connect()
-
-            // Park the structural collector for the ENTIRE flood window: it takes
-            // its first event, then blocks on `floodDrained`. On base this lets
-            // the reader's 256-slot bus buffer fill with %output copies while the
-            // window-close is tryEmit'd into a full buffer (dropped).
-            val collector = scope.async {
-                client.events.collect { ev ->
-                    floodDrained.await()
-                    if (ev is ControlEvent.WindowClose && ev.windowId == "@7") {
-                        windowCloseSeen.complete(ev.windowId)
-                    }
-                }
-            }
-            delay(100) // let the collector subscribe to the hot bus
-
-            // Flood far past EVENT_BUFFER (256), then the trailing window-close.
-            shell.feed(
-                buildString {
-                    repeat(1000) { index ->
-                        append("%output %0 frame-")
-                        append(index)
-                        append('\n')
-                    }
-                    append("%window-close @7\n")
-                },
-            )
-            // Reader-progress barrier that works on BOTH base and fix: a per-pane
-            // sentinel delivered via outputFor (never the lossy events bus). Once
-            // it arrives, the reader has processed PAST the window-close line.
-            shell.feed("%output %barrier done\n")
-            withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { client.outputFor("%barrier").first() }
-
-            // Release the parked collector and drain the buffered events. On base
-            // the window-close is gone (dropped under the flood) → this times out
-            // (RED). With the fix it is delivered (GREEN).
-            floodDrained.complete(Unit)
-            val closed = withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { windowCloseSeen.await() }
-            assertEquals("@7", closed)
-            collector.cancel()
-        } finally {
-            floodDrained.complete(Unit)
-            client.close()
-        }
-    }
-
-    /**
-     * Issue #1204 — the primary suspect for the fragments-over-black class.
-     *
-     * `%output` that tmux emits for a pane BEFORE its per-pane pipe is
-     * registered by [RealTmuxClient.outputFor] used to be discarded silently
-     * (the null-safe `paneOutputPipes[paneId]?.send(event)` no-op) — no
-     * counter, no diagnostic, no recovery. On a Claude pane, whose TUI only
-     * repaints incrementally, the lost first frame stays black indefinitely.
-     *
-     * G10 red→green core: feed two frames for `%0` before anyone registers
-     * `outputFor("%0")`, then register + collect. On base the buffered bytes
-     * are gone (this `take(3)` never completes → timeout); with the bounded
-     * pre-registration replay buffer all three arrive in order, the two
-     * pre-registration frames correctly interleaved ahead of the live one.
-     */
-    @Test
-    fun `pane output arriving before outputFor registration is replayed in order`() = runBlocking {
-        val shell = FakeShell()
-        val session = FakeSession(shell)
-        val client = RealTmuxClient(session, scope)
-        try {
-            client.connect()
-
-            // Issue #1224: `%output` is no longer on the global events bus, so we
-            // synchronise reader progress on a per-pane SENTINEL delivered via
-            // outputFor (the pre-registration replay path) instead. The
-            // single-threaded reader is strictly FIFO, so once the sentinel frame
-            // arrives, both %0 pre-registration frames ahead of it were buffered.
-            // This barrier is load-independent and works on base AND fix.
-            // %output for %0 BEFORE its per-pane pipe is registered, then a
-            // sentinel frame for a different pane.
-            shell.feed("%output %0 before-one\n%output %0 before-two\n%output %barrier ready\n")
-            withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { client.outputFor("%barrier").first() }
-
-            // Now register the pane pipe and collect. On base, before-one/two
-            // are gone (dropped at the no-pipe site) so this never reaches 3
-            // and times out; with the fix all three arrive in order.
-            val paneEvents = scope.async {
-                client.outputFor("%0").take(3).toList()
-            }
-            delay(100) // let the pane collector attach
-            shell.feed("%output %0 after-one\n")
-
-            val events = withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { paneEvents.await() }
-
-            assertEquals(
-                listOf("before-one", "before-two", "after-one"),
-                events.map { String(it.data, StandardCharsets.US_ASCII) },
-            )
-        } finally {
-            client.close()
-        }
-    }
-
-    /**
-     * Issue #1204 — class coverage (G2): the attach/switch re-registration
-     * case. A background/other-window pane (`%1`) floods output while the user
-     * is on `%0`; the user only later switches to `%1` (first
-     * `outputFor("%1")`). The buffered `%1` frames must replay on switch, and
-     * the co-registered `%0` pipe must NOT consume `%1`'s buffer (per-pane
-     * demux is preserved).
-     */
-    @Test
-    fun `switching to a pane replays output buffered before its first registration`() = runBlocking {
-        val shell = FakeShell()
-        val session = FakeSession(shell)
-        val client = RealTmuxClient(session, scope)
-        try {
-            client.connect()
-
-            // %0 is the currently-attached pane.
-            val pane0 = scope.async { client.outputFor("%0").take(1).toList() }
-            delay(100)
-
-            // Issue #1224: `%output` is no longer on the global events bus, so
-            // synchronise reader progress on a per-pane SENTINEL delivered via
-            // outputFor (the FIFO reader guarantees the background %1 frames ahead
-            // of the sentinel were buffered by the time it arrives).
-            shell.feed(
-                "%output %0 live0\n" +
-                    "%output %1 bg-one\n" +
-                    "%output %1 bg-two\n" +
-                    "%output %barrier ready\n",
-            )
-            withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { client.outputFor("%barrier").first() }
-
-            // User switches to %1 — first registration replays the buffered
-            // background frames in order.
-            val pane1 = scope.async { client.outputFor("%1").take(3).toList() }
-            delay(100)
-            shell.feed("%output %1 bg-three\n")
-
-            val p0 = withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { pane0.await() }
-            val p1 = withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { pane1.await() }
-
-            assertEquals(listOf("live0"), p0.map { String(it.data, StandardCharsets.US_ASCII) })
-            assertEquals(
-                listOf("bg-one", "bg-two", "bg-three"),
-                p1.map { String(it.data, StandardCharsets.US_ASCII) },
-            )
-        } finally {
-            client.close()
-        }
-    }
-
-    /**
-     * Issue #1204 — class coverage (G2): the buffer-overflow eviction case.
-     * More pre-registration frames than the cap arrive; the oldest are evicted
-     * (bounded), the drop is counted, and the eviction fires the
-     * `tmux_client_preregistration_output_drop` diagnostic into the exportable
-     * JSONL sink so this loss is never invisible again.
-     */
-    @Test
-    fun `pre-registration buffer overflow evicts oldest and records drop diagnostic`() = runBlocking {
-        val diagnostics = Collections.synchronizedList(
-            mutableListOf<Pair<String, Map<String, Any?>>>(),
-        )
-        TmuxClientDiagnostics.install { event, fields -> diagnostics += event to fields }
-        val shell = FakeShell()
-        val session = FakeSession(shell)
-        val client = RealTmuxClient(session, scope)
-        try {
-            client.connect()
-
-            // 300 > the 256-event pre-registration cap → 44 evictions.
-            val target = 300
-            val cap = 256
-
-            // Feed all 300 `%9` frames, then ONE sentinel frame for a different
-            // pane (`%8`). The tmux control reader processes lines strictly FIFO,
-            // so the sentinel is delivered only after every `%9` frame ahead of
-            // it has been buffered (with eviction). We synchronise on the sentinel
-            // arriving via the *reliable* per-pane pre-registration replay
-            // (`outputFor("%8")`) rather than the lossy global `client.events`
-            // SharedFlow: with replay=0 and a 256-event bus buffer, a flood of 300
-            // frames drops events whenever the bus collector lags under CI load —
-            // `seen` then never reaches 300 and the old `withTimeout` on it expired
-            // with a TimeoutCancellationException (#1252). The pre-registration
-            // buffer never drops the sentinel, so this barrier is load-independent.
-            shell.feed(
-                buildString {
-                    repeat(target) { i ->
-                        append("%output %9 frame-")
-                        append(i)
-                        append('\n')
-                    }
-                    append("%output %8 sentinel\n")
-                },
-            )
-            withTimeout(ASYNC_AWAIT_TIMEOUT_MS) {
-                client.outputFor("%8").first()
-            }
-
-            // Register: only the newest `cap` survive; the oldest were evicted.
-            val events = withTimeout(ASYNC_AWAIT_TIMEOUT_MS) {
-                client.outputFor("%9").take(cap).toList()
-            }
-
-            // Bounded: exactly the cap survives, oldest evicted, order kept.
-            assertEquals(cap, events.size)
-            assertEquals(
-                "frame-${target - cap}",
-                String(events.first().data, StandardCharsets.US_ASCII),
-            )
-            assertEquals(
-                "frame-${target - 1}",
-                String(events.last().data, StandardCharsets.US_ASCII),
-            )
-
-            // Observable: the eviction fired the drop diagnostic into the JSONL.
-            val drop = diagnostics.firstOrNull {
-                it.first == "tmux_client_preregistration_output_drop"
-            }
-            assertTrue("expected a pre-registration drop diagnostic to be recorded", drop != null)
-            assertEquals("%9", drop!!.second["pane"])
-            assertEquals(1, drop.second["droppedEvents"])
-        } finally {
-            TmuxClientDiagnostics.install(TmuxClientDiagnosticSink.Noop)
-            client.close()
-        }
-    }
-
-    /**
-     * Issue #1231 T2 — the LF-framing accumulation buffer in the reader loop
-     * must be bounded. An LF-starved stream (a binary MOTD before the `-CC`
-     * handshake, a degraded non-control-mode byte stream, or a wedged server
-     * that stops emitting LFs) grows the per-line `ByteArrayOutputStream` one
-     * byte at a time. Without a ceiling it grows until the process OOMs, with
-     * NO diagnostic. This test feeds ~1.5 MB with no LF and asserts the buffer
-     * is capped-and-reset (so it never grows to the full feed size), the
-     * `tmux_client_line_overflow` diagnostic fires repeatedly, and normal event
-     * framing resumes once an LF finally arrives.
-     *
-     * Red→green: on base (unbounded buffer) no diagnostic is ever recorded, so
-     * the `overflow.isNotEmpty()` assertion is RED.
-     */
-    @Test
-    fun `LF-starved stream caps the framing buffer and records overflow diagnostic`() = runBlocking {
-        // Mirrors the private TmuxClient.MAX_LINE_BUFFER_BYTES ceiling (512 KB).
-        val maxLineBytes = 512 * 1024
-        val diagnostics = Collections.synchronizedList(
-            mutableListOf<Pair<String, Map<String, Any?>>>(),
-        )
-        TmuxClientDiagnostics.install { event, fields -> diagnostics += event to fields }
-        val shell = FakeShell()
-        val session = FakeSession(shell)
-        val client = RealTmuxClient(session, scope)
-        try {
-            client.connect()
-
-            // Subscribe to the recovered pane's output BEFORE feeding so the
-            // post-overflow frame can't be missed in the subscribe race.
-            val recovered = scope.async {
-                client.outputFor("%0").first()
-            }
-            // Give the collector a tick to subscribe to the per-pane flow.
-            delay(200)
-
-            // Exactly 2× the cap of LF-free bytes → two clean flush-and-reset
-            // overflows that leave the buffer empty, so the trailing frame is
-            // framed cleanly. On the unbounded reader this all accumulates in
-            // ONE ever-growing line buffer that never resets (and would OOM in
-            // production); no overflow diagnostic is ever recorded.
-            val starved = "x".repeat(maxLineBytes * 2)
-            // Then a real, LF-terminated %output frame proves framing resumed.
-            shell.feed(starved + "%output %0 recovered\n")
-
-            // Barrier: the post-overflow frame must reach the pane output.
-            val frame = withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { recovered.await() }
-            assertEquals("recovered", String(frame.data, StandardCharsets.US_ASCII))
-
-            val overflow = diagnostics.filter { it.first == "tmux_client_line_overflow" }
-            assertTrue(
-                "expected at least one tmux_client_line_overflow diagnostic; got none " +
-                    "(unbounded buffer would have OOMed instead of flushing)",
-                overflow.isNotEmpty(),
-            )
-            // Two fires over the 2× feed prove the buffer resets each time
-            // rather than growing unbounded to the full 1 MB.
-            assertTrue(
-                "expected the buffer to cap-and-reset repeatedly, got ${overflow.size} overflow(s)",
-                overflow.size >= 2,
-            )
-            // Each overflow flushed at exactly the ceiling — never the full feed.
-            overflow.forEach { (_, fields) ->
-                assertEquals(maxLineBytes, fields["maxBytes"])
-                assertEquals(maxLineBytes, fields["bytes"])
-            }
-        } finally {
-            TmuxClientDiagnostics.install(TmuxClientDiagnosticSink.Noop)
-            client.close()
-        }
-    }
-
-    // ── Issue #1212: pre-registration buffer lifecycle hardening ──────────────
-
-    /**
-     * Issue #1212 (AC1) — a pane that dies before its pipe is ever registered
-     * must not retain its pre-registration replay buffer for the connection's
-     * lifetime. tmux announces the death via `%window-close @<w>`; the window's
-     * pane set is learned from its `%layout-change` layout, so on window-close
-     * the dead panes' buffers are released.
-     *
-     * Red→green: on base (no cleanup) the `%0` buffer survives window-close and
-     * still replays the doomed frame on a later registration; with the fix the
-     * buffer is released (count + bytes drop to 0) and the doomed frame is gone.
-     */
-    @Test
-    fun `pre-registration buffer is released when its window closes`() = runBlocking {
-        val shell = FakeShell()
-        val session = FakeSession(shell)
-        val client = RealTmuxClient(session, scope)
-        try {
-            client.connect()
-
-            val closeSeen = CompletableDeferred<Unit>()
-            val watcher = scope.async {
-                client.events.collect { ev ->
-                    if (ev is ControlEvent.WindowClose && ev.windowId == "@1") {
-                        closeSeen.complete(Unit)
-                    }
-                }
-            }
-            delay(100)
-
-            // @1 owns pane %0 (a single-pane layout leaf `...,0`). %0 floods
-            // output before anyone registers outputFor("%0"), then @1 closes.
-            // The reader processes these in order on one loop, and window-close
-            // cleanup runs BEFORE the event reaches the bus — so once the
-            // watcher sees WindowClose, cleanup has already happened.
-            shell.feed("%layout-change @1 bffb,80x24,0,0,0\n")
-            shell.feed("%output %0 doomed\n")
-            shell.feed("%window-close @1\n")
-            withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { closeSeen.await() }
-            watcher.cancel()
-
-            // The buffer for %0 is gone (released on window-close).
-            assertEquals(0, client.preRegistrationBufferCountForTest())
-            assertEquals(0L, client.preRegistrationRetainedBytesForTest())
-
-            // Behavioural: registering %0 now replays nothing stale — the first
-            // frame the collector sees is the fresh live one, not "doomed".
-            val pane0 = scope.async { client.outputFor("%0").take(1).toList() }
-            delay(100)
-            shell.feed("%output %0 live\n")
-            val got = withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { pane0.await() }
-            assertEquals(
-                listOf("live"),
-                got.map { String(it.data, StandardCharsets.US_ASCII) },
-            )
-        } finally {
-            client.close()
-        }
-    }
-
-    /**
-     * Issue #1212 (AC2) — the TOTAL retained pre-registration bytes and the
-     * number of distinct retained buffers are bounded across MANY panes, not
-     * just per pane. Without a global cap, N never-registered background panes
-     * each pin a full per-pane buffer (256 KB × N). Here 200 orphaned panes
-     * each flood 20 KB (4 MB attempted); the global caps hold the retained
-     * total under the byte cap AND the distinct-pane cap.
-     *
-     * Red→green: on base (no global cap) retained bytes ≈ 4 MB and count = 200;
-     * with the fix count ≤ the pane cap and bytes ≤ the total-byte cap.
-     */
-    @Test
-    fun `many orphaned panes are bounded by the global pre-registration caps`() = runBlocking {
-        val shell = FakeShell()
-        val session = FakeSession(shell)
-        val client = RealTmuxClient(session, scope)
-        try {
-            client.connect()
-
-            val paneCount = 200
-            val frameBytes = 20 * 1024
-            val payload = "x".repeat(frameBytes)
-
-            for (i in 0 until paneCount) {
-                shell.feed("%output %p$i $payload\n")
-            }
-            // A sentinel pane fed LAST: the single-threaded reader processes it
-            // only after every fat frame before it. Issue #1224: `%output` is no
-            // longer on the events bus, so we synchronise on the sentinel's
-            // per-pane delivery via outputFor (the FIFO reader guarantees all 200
-            // panes' output was processed by the time it arrives).
-            shell.feed("%output %sentinel done\n")
-            withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { client.outputFor("%sentinel").first() }
-
-            val retainedPanes = client.preRegistrationBufferCountForTest()
-            val retainedBytes = client.preRegistrationRetainedBytesForTest()
-
-            // Bounded distinct-pane count (+1 slack for the tiny sentinel pane).
-            assertTrue(
-                "retained pane buffers ($retainedPanes) must be bounded well below the $paneCount fed",
-                retainedPanes <= 64 + 1,
-            )
-            // Bounded aggregate bytes: far below the ~4 MB that would accrue
-            // unbounded, held at/under the 1 MB global byte cap.
-            assertTrue(
-                "retained bytes ($retainedBytes) must be bounded by the global byte cap",
-                retainedBytes <= 1024L * 1024L,
-            )
-        } finally {
-            client.close()
-        }
-    }
-
-    /**
-     * Issue #1212 (AC3) — a pane pipe registered via [TmuxClient.outputFor]
-     * whose returned flow is NEVER collected must not wedge the client or pin
-     * its replay for the connection's lifetime. The replay job waits a bounded
-     * grace for the first subscriber, then abandons (releases) the replay and
-     * proceeds — so the client keeps serving other panes.
-     *
-     * Red→green: on base the job parks forever on `subscriptionCount.first{>0}`
-     * and the abandon diagnostic never fires (the wait below times out); with
-     * the fix the grace elapses, the diagnostic fires, and a different pane
-     * still works end-to-end.
-     */
-    @Test
-    fun `registered pane whose flow is never collected releases its replay and does not wedge`() = runBlocking {
-        val diagnostics = Collections.synchronizedList(
-            mutableListOf<Pair<String, Map<String, Any?>>>(),
-        )
-        TmuxClientDiagnostics.install { event, fields -> diagnostics += event to fields }
-        val shell = FakeShell()
-        val session = FakeSession(shell)
-        val client = RealTmuxClient(session, scope, firstSubscriberReplayGraceMs = 200L)
-        try {
-            client.connect()
-
-            // Buffer a pre-registration frame for %0 before it registers. Issue
-            // #1224: `%output` is no longer on the events bus, so synchronise on a
-            // per-pane SENTINEL (delivered via outputFor) — the FIFO reader
-            // guarantees %0's frame was buffered by the time the sentinel arrives.
-            shell.feed("%output %0 buffered\n%output %barrier ready\n")
-            withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { client.outputFor("%barrier").first() }
-
-            // Register the pipe but NEVER collect the returned flow.
-            client.outputFor("%0")
-
-            // After the grace window with no collector, the replay is abandoned.
-            withTimeout(ASYNC_AWAIT_TIMEOUT_MS) {
-                while (diagnostics.none { it.first == "tmux_client_preregistration_replay_abandoned" }) {
-                    delay(20)
-                }
-            }
-            val abandoned = diagnostics.first {
-                it.first == "tmux_client_preregistration_replay_abandoned"
-            }
-            assertEquals("%0", abandoned.second["pane"])
-
-            // Not wedged: a DIFFERENT pane still delivers output end-to-end.
-            val pane1 = scope.async { client.outputFor("%1").take(1).toList() }
-            delay(100)
-            shell.feed("%output %1 live1\n")
-            val got = withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { pane1.await() }
-            assertEquals(
-                listOf("live1"),
-                got.map { String(it.data, StandardCharsets.US_ASCII) },
-            )
-        } finally {
-            TmuxClientDiagnostics.install(TmuxClientDiagnosticSink.Noop)
-            client.close()
-        }
-    }
-
-    @Test
-    fun `codex scale output flood cannot starve command response parsing`() = runBlocking {
-        val shell = FakeShell()
-        val session = FakeSession(shell)
-        val client = RealTmuxClient(session, scope, commandTimeoutMs = 1_000L)
-        val outputCount = 900
-        val firstOutputBlocked = CompletableDeferred<Unit>()
-        val releasePaneCollector = CompletableDeferred<Unit>()
-        val allOutputDelivered = CompletableDeferred<Unit>()
-        val delivered = AtomicInteger(0)
-        val firstPayload = arrayOfNulls<String>(1)
-        val lastPayload = arrayOfNulls<String>(1)
-        try {
-            client.connect()
-            withTimeout(2_000) {
-                while (shell.stdinBytes().isEmpty()) { yield(); delay(10) }
-            }
-            shell.resetStdin()
-
-            val slowPaneCollector = scope.async {
-                client.outputFor("%0").collect {
-                    firstOutputBlocked.complete(Unit)
-                    releasePaneCollector.await()
-                    val text = String(it.data, StandardCharsets.UTF_8)
-                    val index = delivered.incrementAndGet()
-                    if (index == 1) firstPayload[0] = text
-                    if (index == outputCount) {
-                        lastPayload[0] = text
-                        allOutputDelivered.complete(Unit)
-                    }
-                }
-            }
-            delay(100)
-
-            val response = scope.async {
-                client.sendCommand("display-message -p ok")
-            }
-            withTimeout(2_000) {
-                while (!shell.stdinAsString().endsWith("display-message -p ok\n")) {
-                    yield(); delay(10)
-                }
-            }
-
-            val feedJob = scope.async {
-                shell.feed(codexScaleControlModeFlood(commandNumber = 1L, outputCount = outputCount))
-            }
-
-            withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { firstOutputBlocked.await() }
-            val result = withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { response.await() }
-            assertFalse(result.isError)
-            assertEquals(listOf("ok"), result.output)
-            assertFalse(
-                "slow terminal output fanout must not be classified as a tmux disconnect",
-                client.disconnected.value,
-            )
-            withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { feedJob.await() }
-
-            releasePaneCollector.complete(Unit)
-            withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { allOutputDelivered.await() }
-            assertEquals(
-                "pane output backlog must remain lossless; do not silently drop terminal bytes",
-                outputCount,
-                delivered.get(),
-            )
-            assertTrue(
-                "first output should be preserved",
-                firstPayload[0]?.contains("codex-flood-0000") == true,
-            )
-            assertTrue(
-                "last output should be preserved",
-                lastPayload[0]?.contains("codex-flood-0899") == true,
-            )
-            slowPaneCollector.cancel()
-        } finally {
-            releasePaneCollector.complete(Unit)
-            client.close()
-        }
-    }
-
-    @Test
-    fun `unbounded pane output flood emits overflow without disconnecting reader`() = runBlocking {
-        val shell = FakeShell()
-        val session = FakeSession(shell)
-        val client = RealTmuxClient(session, scope, commandTimeoutMs = 1_000L)
-        val firstOutputBlocked = CompletableDeferred<Unit>()
-        val releasePaneCollector = CompletableDeferred<Unit>()
-        val diagnosticEvents = Collections.synchronizedList(
-            mutableListOf<Pair<String, Map<String, Any?>>>(),
-        )
-        installDiagnosticsForClient(
-            client,
-            setOf("tmux_client_pane_output_backlog_overflow"),
-            diagnosticEvents,
-        )
-        try {
-            client.connect()
-            withTimeout(2_000) {
-                while (shell.stdinBytes().isEmpty()) { yield(); delay(10) }
-            }
-            shell.resetStdin()
-
-            val blockedPaneCollector = scope.async {
-                client.outputFor("%0").collect {
-                    firstOutputBlocked.complete(Unit)
-                    releasePaneCollector.await()
-                }
-            }
-            val overflow = scope.async {
-                client.outputBacklogOverflows.first()
-            }
-            delay(100)
-
-            val response = scope.async {
-                client.sendCommand("display-message -p ok")
-            }
-            withTimeout(2_000) {
-                while (!shell.stdinAsString().endsWith("display-message -p ok\n")) {
-                    yield(); delay(10)
-                }
-            }
-
-            val feedJob = scope.async {
-                shell.feed(codexScaleControlModeFlood(commandNumber = 1L, outputCount = 5_000))
-            }
-
-            withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { firstOutputBlocked.await() }
-            val overflowEvent = withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { overflow.await() }
-            assertEquals("%0", overflowEvent.paneId)
-            assertTrue("overflow must report dropped events", overflowEvent.droppedEvents > 0)
-            withTimeout(ASYNC_AWAIT_TIMEOUT_MS) {
-                while (diagnosticEvents.isEmpty()) { yield(); delay(10) }
-            }
-
-            val result = withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { response.await() }
-            assertFalse(result.isError)
-            assertEquals(listOf("ok"), result.output)
-            assertFalse(
-                "output backlog overflow is a local terminal condition, not a transport disconnect",
-                client.disconnected.value,
-            )
-            withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { feedJob.await() }
-            assertEquals(
-                "core overflow diagnostics must be first-overflow-only per pipe",
-                1,
-                diagnosticEvents.size,
-            )
-            val fields = diagnosticEvents.single().second
-            assertEquals("%0", fields["pane"])
-            assertEquals("pocketshell", fields["session"])
-            assertTrue("overflow diagnostics should include stable client id", fields["clientId"] is Long)
-            assertTrue("overflow diagnostics should include client hash", fields["clientHash"] is Int)
-            assertEquals(4096, fields["capacity"])
-            blockedPaneCollector.cancel()
-        } finally {
-            releasePaneCollector.complete(Unit)
-            client.close()
-        }
-    }
-
-    /**
-     * Issue #1205 — the CORE reproduction of the trySend-drop that used to KILL
-     * the pane. A sustained %output flood past the bounded [OUTPUT_BACKLOG_EVENTS]
-     * channel (with the pane collector parked so the drain job can't keep up)
-     * makes `trySend` drop and fires [TmuxClient.outputBacklogOverflows] — exactly
-     * the signal the app latched `surfaceError` on. The FIX's recovery path drains
-     * the pane's queued backlog before a `capture-pane` reseed so the stale burst
-     * frames can't replay on top of the authoritative snapshot; this proves the
-     * NEW [TmuxClient.drainPaneOutputBacklog] actually empties the saturated
-     * channel (returns > 0 for the overflowed pane) and is a no-op for an
-     * unregistered pane.
-     */
-    @Test
-    fun `drainPaneOutputBacklog empties a saturated pane channel so a reseed is authoritative`() =
-        runBlocking {
-            val shell = FakeShell()
-            val session = FakeSession(shell)
-            val client = RealTmuxClient(session, scope, commandTimeoutMs = 1_000L)
-            val firstOutputBlocked = CompletableDeferred<Unit>()
-            val releasePaneCollector = CompletableDeferred<Unit>()
-            try {
-                client.connect()
-                withTimeout(2_000) {
-                    while (shell.stdinBytes().isEmpty()) { yield(); delay(10) }
-                }
-                shell.resetStdin()
-
-                // Park the pane collector on the first frame: the drain job stalls,
-                // the flow buffer fills, the channel backs up, and the flood below
-                // overflows it (the exact trySend-drop condition).
-                val blockedPaneCollector = scope.async {
-                    client.outputFor("%0").collect {
-                        firstOutputBlocked.complete(Unit)
-                        releasePaneCollector.await()
-                    }
-                }
-                val overflow = scope.async { client.outputBacklogOverflows.first() }
-                delay(100)
-
-                val feedJob = scope.async {
-                    shell.feed(codexScaleControlModeFlood(commandNumber = 1L, outputCount = 5_000))
-                }
-
-                withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { firstOutputBlocked.await() }
-                val overflowEvent = withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { overflow.await() }
-                assertEquals("%0", overflowEvent.paneId)
-                assertTrue("sustained flood must overflow the channel", overflowEvent.droppedEvents > 0)
-                withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { feedJob.await() }
-
-                // LOAD-BEARING: the queued stale frames ARE drainable — the recovery
-                // clears them before the capture-pane reseed so they cannot double
-                // apply on top of the authoritative grid. A saturated channel has
-                // frames to drop.
-                val drained = client.drainPaneOutputBacklog("%0")
-                assertTrue(
-                    "drainPaneOutputBacklog must empty the saturated backlog (drained=$drained)",
-                    drained > 0,
-                )
-                // Draining again is now a no-op (already emptied), and an
-                // unregistered pane returns 0 without throwing.
-                assertEquals(0, client.drainPaneOutputBacklog("%0"))
-                assertEquals(0, client.drainPaneOutputBacklog("%does-not-exist"))
-                assertFalse(
-                    "draining the backlog is a local recovery, never a transport disconnect",
-                    client.disconnected.value,
-                )
-
-                blockedPaneCollector.cancel()
-            } finally {
-                releasePaneCollector.complete(Unit)
-                client.close()
-            }
-        }
 
     @Test
     fun `close fails an in-flight sendCommand with TmuxClientException`() = runBlocking {
@@ -2946,77 +1900,6 @@ class TmuxClientTest {
             before,
             after,
         )
-    }
-
-    @Test
-    fun `output events are not lost between begin and end response framing`() = runBlocking {
-        // tmux can interleave `%output` (a notification) with response
-        // payload as long as it's not inside the response block. We make
-        // sure ordinary pane output still flows through cleanly after a
-        // sendCommand cycle. Issue #1224: `%output` is delivered via the
-        // per-pane [outputFor] pipe (no longer the structural events bus), so we
-        // assert both the pre-block and post-block frames land on %0's stream.
-        val shell = FakeShell()
-        val session = FakeSession(shell)
-        val client = RealTmuxClient(session, scope)
-        try {
-            client.connect()
-            withTimeout(2_000) {
-                while (shell.stdinBytes().isEmpty()) { yield(); delay(10) }
-            }
-
-            val collected = scope.async {
-                client.outputFor("%0").take(2).toList()
-            }
-            delay(100)
-
-            val response = scope.async { client.sendCommand("list-sessions") }
-            withTimeout(2_000) {
-                while (!shell.stdinAsString().endsWith("list-sessions\n")) {
-                    yield(); delay(10)
-                }
-            }
-            shell.feed(
-                "%output %0 before\n" +
-                    "%begin 1 1 0\n" +
-                    "row\n" +
-                    "%end 1 1 0\n" +
-                    "%output %0 after\n",
-            )
-            val r = withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { response.await() }
-            assertEquals(listOf("row"), r.output)
-            val events = withTimeout(ASYNC_AWAIT_TIMEOUT_MS) { collected.await() }
-            assertEquals(2, events.size)
-            assertEquals("before", String(events[0].data, StandardCharsets.US_ASCII))
-            assertEquals("after", String(events[1].data, StandardCharsets.US_ASCII))
-        } finally {
-            client.close()
-        }
-    }
-
-    private fun codexScaleControlModeFlood(commandNumber: Long, outputCount: Int): String = buildString {
-        repeat(outputCount) { index ->
-            append("%output %0 ")
-            append("\\033[38;5;")
-            append(index % 256)
-            append('m')
-            append("codex-flood-")
-            append(index.toString().padStart(4, '0'))
-            append(' ')
-            append("x".repeat(220))
-            if (index % 5 == 0) append("\\033[2K")
-            if (index % 11 == 0) append("\\rspinner-frame-")
-            append(index)
-            append("\\033[0m")
-            append('\n')
-        }
-        append("%begin 1 ")
-        append(commandNumber)
-        append(" 0\n")
-        append("ok\n")
-        append("%end 1 ")
-        append(commandNumber)
-        append(" 0\n")
     }
 
     private suspend fun assertBestEffortTimeoutDrainsLateResponse(
